removed header super component
