+++ src/__mocks__/useStaticQuery.js
@@ -1,0 +1,16 @@
+import * as Gatsby from 'gatsby';
+
+export const useStaticQuery = jest.spyOn(Gatsby, `useStaticQuery`);
+export const mockUseStaticQuery = {
+  site: {
+    siteMetadata: {
+      metatags: {
+        title: `Gatsby Default Starter`,
+        author: 'Diego Costa',
+      },
+    },
+  },
+  pages: {
+    nodes: [{}],
+  },
+};
