+++ src/components/Divisor.tsx
