+++ src/components/Footer.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 
 import { useLocale } from '~/hooks/useLocale';
 
-const Content = styled.footer`
+const Container = styled.footer`
   margin-bottom: 50px;
 
   nav {
@@ -38,23 +38,21 @@ const Link = styled.a.attrs({
 `;
 
 interface FooterProps {
-  year: number;
   sourceCode?: string;
   author?: string;
 }
 
 export default function Footer(props: FooterProps): React.ReactElement {
   const locale = useLocale();
+  const year = new Date().getFullYear();
 
   return (
-    <>
-      <Content>
-        <Label>
-          CC-BY {props.year} {props.author}, {locale.getTranslationFor('footer.builtWith')}{' '}
-          <Link href="https://gatsbyjs.org">gatsby</Link> •{' '}
-          {props.sourceCode && <Link href={props.sourceCode}>{locale.getTranslationFor('footer.sourceCode')}</Link>}
-        </Label>
-      </Content>
-    </>
+    <Container>
+      <Label>
+        CC-BY {year} {props.author}, {locale.getTranslationFor('footer.builtWith')}{' '}
+        <Link href="https://gatsbyjs.org">gatsby</Link> •{' '}
+        {props.sourceCode && <Link href={props.sourceCode}>{locale.getTranslationFor('footer.sourceCode')}</Link>}
+      </Label>
+    </Container>
   );
 }
