--- src/components/Header/components/LanguageSwitcher/index.tsx
@@ -1,15 +1,0 @@
-import { graphql } from 'gatsby';
-
-export { default } from './LanguageSwitcher';
-
-export const query = graphql`
-  fragment LanguageInformation on LocaleConnection {
-    edges {
-      node {
-        ns
-        data
-        language
-      }
-    }
-  }
-`;
