--- src/components/Header/components/ThemeSwitcher/index.ts
@@ -1,1 +1,0 @@
-export { default } from './ThemeSwitcher';
