+++ src/components/__tests__/Footer.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 
-import Footer from '../Footer';
+import Footer from '~/components/Footer';
 
 describe('<Footer />', () => {
   it('renders properly', () => {
