+++ src/components/__tests__/ThemeSwitcher.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 
-import ThemeSwitcher from '../ThemeSwitcher';
+import ThemeSwitcher from '~/components/ThemeSwitcher';
 
 describe('<ThemeSwitcher />', () => {
   it('renders properly', () => {
