+++ src/pages/__tests__/404.test.js
@@ -1,12 +1,20 @@
 import React from 'react';
 import { render } from '@testing-library/react';
 
+import { useStaticQuery, mockUseStaticQuery } from '../../__mocks__/useStaticQuery';
 import NotFoundPage from '../404';
 
-jest.mock('../../components/Header', () => () => <p>Header</p>);
 jest.mock('../../components/Footer', () => () => <p>Footer</p>);
 
 describe('<NotFoundPage>', () => {
+  beforeEach(() => {
+    useStaticQuery.mockImplementation(() => mockUseStaticQuery);
+  });
+
+  afterEach(() => {
+    jest.restoreAllMocks();
+  });
+
   it('renders properly', () => {
     const { baseElement } = render(<NotFoundPage />);
     expect(baseElement).toMatchSnapshot();
