+++ src/pages/links.tsx
@@ -1,1 +1,11 @@
+import { graphql } from 'gatsby';
+
 export { default } from '~/templates/Links';
+
+export const pageQuery = graphql`
+  query LinksPage($language: String!) {
+    locales: allLocale(filter: { language: { eq: $language } }) {
+      ...LanguageInformation
+    }
+  }
+`;
